Long overdue splitting of export GMDSceneGatherers into different files
